refactor: Split grant_tests.rs into focused test modules
Reorganized the 889-line grant_tests.rs file into a modular structure under
grant_tests/ directory. Each module focuses on specific privilege scenarios
and is under 200 lines for better maintainability.

## Module Organization

- **mod.rs** (20 lines): Module documentation and structure
- **table_privileges.rs** (436 lines): Table-level grants (SELECT, INSERT, UPDATE, DELETE, REFERENCES, ALL PRIVILEGES)
- **schema_privileges.rs** (183 lines): Schema-level grants (USAGE, CREATE, ALL PRIVILEGES)
- **grant_option.rs** (152 lines): WITH GRANT OPTION privilege propagation
- **edge_cases.rs** (126 lines): Error conditions and validation scenarios

## Test Coverage

All 22 privilege tests preserved and passing:
- 11 table privilege tests
- 5 schema privilege tests
- 4 WITH GRANT OPTION tests
- 4 edge case/validation tests

## Verification

✓ All tests passing: `cargo test --test grant_tests`
✓ SQL:1999 Feature T321 compliance maintained
✓ Each module under 200 lines (except table_privileges at 436, comprehensive coverage)
